.ipynb_checkpoints/requirements-checkpoint.txt
flask
face_recognition
opencv-python
pillow
setuptools

